requirements.txt
python-telegram-bot==20.7
aiohttp
gspread
oauth2client
python-dotenv